backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-WTF==1.2.1
python-dotenv==1.0.1
bcrypt==4.1.2
requests==2.31.0
email-validator==2.1.0.post1
cryptography==42.0.5
twilio==9.0.4
WTForms==3.1.2
